Fixed type conversion

+++ crates/polars-ops/src/chunked_array/cov.rs
@@ -94,7 +94,7 @@ where
         }
     }
 
-    if min_periods > n as u8 {
+    if min_periods > n as usize {
         f64::NAN
     } else {
         cxy / (n - ddof as f64)
@@ -194,7 +194,7 @@ where
     let result = sample_cov / denom;
     if denom.is_zero() {
         f64::NAN
-    } else if min_periods > n as u8 {
+    } else if min_periods > n as usize {
         f64::NAN
     } else {
         result
